refactor: Move json loading to providers from MapFeatureService
