+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/PlaceListProvider.java
@@ -4,6 +4,10 @@
  */
 package com.wynntils.services.mapdata.providers.builtin;
 
+import com.google.gson.reflect.TypeToken;
+import com.wynntils.core.components.Managers;
+import com.wynntils.core.net.Download;
+import com.wynntils.core.net.UrlId;
 import com.wynntils.services.mapdata.features.PlaceLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
 import java.util.ArrayList;
@@ -12,6 +16,10 @@ import java.util.stream.Stream;
 
 public class PlaceListProvider extends BuiltInProvider {
     private static final List<MapFeature> PROVIDED_FEATURES = new ArrayList<>();
+
+    public PlaceListProvider() {
+        loadPlaces();
+    }
 
     @Override
     public String getProviderId() {
@@ -23,13 +31,14 @@ public class PlaceListProvider extends BuiltInProvider {
         return PROVIDED_FEATURES.stream();
     }
 
-    public void updatePlaces(List<PlaceLocation> places) {
-        PROVIDED_FEATURES.forEach(this::notifyCallbacks);
-        PROVIDED_FEATURES.clear();
-        places.forEach(PlaceListProvider::registerFeatures);
-    }
-
-    private static void registerFeatures(PlaceLocation location) {
-        PROVIDED_FEATURES.add(location);
+    private void loadPlaces() {
+        Download dl = Managers.Net.download(UrlId.DATA_STATIC_PLACE_MAPFEATURES);
+        dl.handleReader(reader -> {
+            TypeToken<List<PlaceLocation>> type = new TypeToken<>() {};
+            List<PlaceLocation> places = Managers.Json.GSON.fromJson(reader, type.getType());
+            PROVIDED_FEATURES.forEach(this::notifyCallbacks);
+            PROVIDED_FEATURES.clear();
+            PROVIDED_FEATURES.addAll(places);
+        });
     }
 }
