+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/ServiceListProvider.java
@@ -4,6 +4,10 @@
  */
 package com.wynntils.services.mapdata.providers.builtin;
 
+import com.google.gson.reflect.TypeToken;
+import com.wynntils.core.components.Managers;
+import com.wynntils.core.net.Download;
+import com.wynntils.core.net.UrlId;
 import com.wynntils.services.mapdata.features.ServiceLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
 import java.util.ArrayList;
@@ -12,6 +16,10 @@ import java.util.stream.Stream;
 
 public class ServiceListProvider extends BuiltInProvider {
     private static final List<MapFeature> PROVIDED_FEATURES = new ArrayList<>();
+
+    public ServiceListProvider() {
+        loadServices();
+    }
 
     @Override
     public String getProviderId() {
@@ -23,13 +31,14 @@ public class ServiceListProvider extends BuiltInProvider {
         return PROVIDED_FEATURES.stream();
     }
 
-    public void updateServices(List<ServiceLocation> services) {
-        PROVIDED_FEATURES.forEach(this::notifyCallbacks);
-        PROVIDED_FEATURES.clear();
-        services.forEach(ServiceListProvider::registerFeature);
-    }
-
-    private static void registerFeature(ServiceLocation location) {
-        PROVIDED_FEATURES.add(location);
+    private void loadServices() {
+        Download dl = Managers.Net.download(UrlId.DATA_STATIC_SERVICE_MAPFEATURES);
+        dl.handleReader(reader -> {
+            TypeToken<List<ServiceLocation>> type = new TypeToken<>() {};
+            List<ServiceLocation> services = Managers.Json.GSON.fromJson(reader, type.getType());
+            PROVIDED_FEATURES.forEach(this::notifyCallbacks);
+            PROVIDED_FEATURES.clear();
+            PROVIDED_FEATURES.addAll(services);
+        });
     }
 }
